requirements.txt
aiogram==3.0.0b7
aiohttp==3.9.3
python-dotenv==1.0.0